Faltan Corregir Todo pero TODO y pulir probar y cosillas varias

+++ app/src/main/java/com/proyecto/gmwork/proyectoandroid/controller/dao/PedidoDAOController.java
@@ -49,7 +49,7 @@ public class PedidoDAOController {
         return client;
     }
 
-    public void removePedido(int id) throws SQLException {
+    public void removePedido(long id) throws SQLException {
         daoPe.delete(daoPe.queryForEq("id", id));
     }
